Refactoring the entire interface
