+++ index.html
@@ -30,10 +30,6 @@
       </button>
     </a>
 
-    <button id="refresh-database-btn" class="nonselectable-text">
-      Update Database
-    </button>
-
     <div id="refresh-wheel" hidden></div>
     <span style="color: var(--button-active); font:bold" id="refresh-status" hidden>
     </span>
@@ -141,9 +137,10 @@
             <button id="search-btn"> Search </button>
           </div>
 
-          <div id="chant-list database" class="section">
+          <!-- Temporary disabled due to architecture updates -->
+          <div id="chant-list database" class="section" hidden>
             <h2> Database </h2>
-            <button id="view-database-btn" class="nonselectable-text"> View database </button>
+            <button id="view-database-btn" class="nonselectable-text" disabled> Browse database </button>
             <div>
               <ul id="database-list">
                 <!-- Chant list goes here -->
