+++ src/index.js
@@ -1,7 +1,7 @@
 import {
     viewDatabaseButton,
     searchModeButton, crossComparisonModeButton,
-    refreshDatabaseButton, refreshStatus, clientVersion, refreshWheel,
+    refreshStatus, clientVersion, refreshWheel,
     databaseList,
     aquitanianCheckbox, squareCheckbox,
     modeCheckboxes, allModeCheckbox, undetectedCheckbox,
@@ -15,7 +15,7 @@ import {
 } from './DOMelements.mjs';
 import {
     drawSVGFromMEIContent, loadMEIFile,
-    checkPersistanceExists, persist, retrieve, getDatabase, env
+    checkPersistanceExists, persist, retrieve, env
 } from './utility/utils.js';
 import {
     clearSearchResultsAndInfo,
@@ -89,20 +89,18 @@ document.addEventListener("DOMContentLoaded", async () => {
         .then(response => response.json());
 
     loadPersistedSearchOptions();
-    let remoteVersion = packageJSON.version;
-    let localVersion = retrieve('version');
-    console.log(`Newest version: ${remoteVersion}, Local version: ${localVersion}`);
-
-    if (localVersion === null || localVersion !== remoteVersion) {
-        await loadDatabaseToLocalStorage();
-        localVersion = remoteVersion;
-    }
+    const remoteVersion = packageJSON.version;
 
     // Display the client version
-    clientVersion.textContent = `Client version: ${localVersion} | Remote version: ${remoteVersion}`;
+    clientVersion.textContent = `Version: ${remoteVersion}`;
 
     if (env === 'development') {
-        clientVersion.textContent += '- Development';
+        clientVersion.textContent += " (Development)";
+    }
+
+    // Clear the chant list if the version is less than 0.3.0
+    if (retrieve('version') !== remoteVersion) {
+        localStorage.removeItem('chantList');
     }
 });
 
@@ -117,94 +115,41 @@ crossComparisonModeButton.addEventListener("click", () => {
     document.getElementById('cross-comparison-panel').hidden = false;
 });
 
-refreshDatabaseButton.addEventListener("click", async () => {
-    clearSearchResultsAndInfo();
-    await loadDatabaseToLocalStorage();
-    if (databaseIsOpen) constructDatabaseList();
-});
-
 /* --------------------- DATABASE --------------------- */
-/**
- * @async
- * @listens viewDatabaseButton
- * Displaying the database as a list of chants on the screen
- */
-async function constructDatabaseList() {
-    /** @type {Chant[]} */
-    const chantList = retrieve('chantList');
-    databaseList.innerHTML = '';
-    for (let chant of chantList) {
-        let li = document.createElement('li');
-        // Let the options contains the title and the notation type of the chant
-        li.textContent = chant.title + " (" + chant.notationType + ")";
-        li.style.wordBreak = "break-word";
-        li.style.cursor = "pointer";
-        li.style.padding = "0.3em 0";
-        li.addEventListener("mouseover", () => {
-            li.style.backgroundColor = "var(--background-hover)";
-        });
-        li.addEventListener("mouseout", () => {
-            li.style.backgroundColor = "white";
-        });
-        li.addEventListener("click", () => {
-            // Scroll to the chant view of the page, smoothly
-            chantDisplay.scrollIntoView({ behavior: "smooth" });
-            // Set the box for the chant
-            chantSVG.style.boxShadow = "0 0 2px 3px #888";
-            chantSVG.innerHTML = drawSVGFromMEIContent(chant.meiContent);
-            // Display the chant information (file name, notation type, mode, etc.)
-            printChantInformation(chant);
-        });
-        databaseList.appendChild(li);
-    }
-}
-
-async function loadDatabaseToLocalStorage() {
-    const database = await getDatabase();
-    console.log(database);
-
-    /** @type {Chant[]} A list of all the chants in the database */
-    let chantList = [];
-    const remoteDatabaseVersion = await fetch("https://raw.githubusercontent.com/ECHOES-from-the-Past/mei-analyser/main/package.json")
-        .then(response => response.json())
-        .then(json => json.version);
-
-    // display the indicator
-    refreshStatus.textContent = `Updating client from version ${retrieve('version')} to ${remoteDatabaseVersion}`;
-    refreshStatus.hidden = false;
-    refreshWheel.hidden = false;
-
-    // Clear the search result display
-    searchResultDiv.innerHTML = '<p> Search results will display here. </p>';
-
-    // Disable the search button
-    searchButton.disabled = true;
-    try {
-        for (let filename of database) {
-            const filePath = rootPath + filename;
-            let MEIFileContentString = await loadMEIFile(filePath);
-            let chant = new Chant(MEIFileContentString, filePath);
-            chantList.push(chant);
-        }
-        persist('chantList', chantList);
-        persist('version', remoteDatabaseVersion);
-    } catch (error) {
-        console.error(error);
-        refreshStatus.textContent = "Error loading database! Please reload the page or report the issue to the developer.";
-        localStorage.removeItem('chantList');
-        refreshWheel.hidden = true;
-        return;
-    }
-
-    refreshStatus.textContent = "Database refresh successfully!";
-    refreshWheel.hidden = true;
-
-    // Enable the search button
-    searchButton.disabled = false;
-    // sleep for 2 seconds
-    await new Promise(resolve => setTimeout(resolve, 2000));
-    refreshStatus.hidden = true;
-}
+// /**
+//  * @async
+//  * @listens viewDatabaseButton
+//  * Displaying the database as a list of chants on the screen
+//  */
+// async function constructDatabaseList() {
+//     /** @type {Chant[]} */
+//     const chantList = retrieve('chantList');
+//     databaseList.innerHTML = '';
+//     for (let chant of chantList) {
+//         let li = document.createElement('li');
+//         // Let the options contains the title and the notation type of the chant
+//         li.textContent = chant.title + " (" + chant.notationType + ")";
+//         li.style.wordBreak = "break-word";
+//         li.style.cursor = "pointer";
+//         li.style.padding = "0.3em 0";
+//         li.addEventListener("mouseover", () => {
+//             li.style.backgroundColor = "var(--background-hover)";
+//         });
+//         li.addEventListener("mouseout", () => {
+//             li.style.backgroundColor = "white";
+//         });
+//         li.addEventListener("click", () => {
+//             // Scroll to the chant view of the page, smoothly
+//             chantDisplay.scrollIntoView({ behavior: "smooth" });
+//             // Set the box for the chant
+//             chantSVG.style.boxShadow = "0 0 2px 3px #888";
+//             chantSVG.innerHTML = drawSVGFromMEIContent(chant.meiContent);
+//             // Display the chant information (file name, notation type, mode, etc.)
+//             printChantInformation(chant);
+//         });
+//         databaseList.appendChild(li);
+//     }
+// }
 
 /* --------------- SEARCH PANEL PERSISTANCE --------------- */
 aquitanianCheckbox.addEventListener("change", () => {
@@ -285,12 +230,16 @@ viewDatabaseButton.addEventListener("click", () => {
 });
 
 /* --------------- SEARCH PANEL --------------- */
-searchButton.addEventListener("click", () => {
+searchButton.addEventListener("click", async () => {
     clearSearchResultsAndInfo();
 
     // Perform search and display the result
-    let searchResults = performSearch();
-    showSearchResult(searchResults);
+    refreshWheel.hidden = false;
+    let searchResults = await performSearch().then((results) => {
+        refreshWheel.hidden = true;
+        showSearchResult(results);
+        return results;
+    });
 });
 
 melismaIncrement.addEventListener("click", () => {
